Implements part of #51
1. Refactored the internal package into a single module instead of sub modules.
2. Started the transition of cli run functions into their proper packages.
3. Organized some of the benchmark module functions into their proper places.
4. Transitioned RunBenchmarks into the benchmark package, implementation still missing!!!
