+++ cli/helpers.go
@@ -5,16 +5,14 @@ package cli
 // on their respective packages.
 
 import (
-	"errors"
 	"fmt"
 	"log/slog"
 
 	"github.com/AlexsanderHamir/prof/engine/benchmark"
 	"github.com/AlexsanderHamir/prof/engine/collector"
+
 	"github.com/AlexsanderHamir/prof/engine/tracker"
-	"github.com/AlexsanderHamir/prof/internal/args"
-	"github.com/AlexsanderHamir/prof/internal/config"
-	"github.com/AlexsanderHamir/prof/internal/shared"
+	"github.com/AlexsanderHamir/prof/internal"
 	"github.com/spf13/cobra"
 )
 
@@ -53,10 +51,10 @@ func CreateRootCmd() *cobra.Command {
 
 func createProfManual() *cobra.Command {
 	manualCmd := &cobra.Command{
-		Use:     shared.MANUALCMD,
+		Use:     internal.MANUALCMD,
 		Short:   "Receives profile files and performs data collection and organization. (doesn't wrap go test)",
 		Args:    cobra.MinimumNArgs(1),
-		Example: fmt.Sprintf("prof %s --tag tagName cpu.prof memory.prof block.prof mutex.prof", shared.MANUALCMD),
+		Example: fmt.Sprintf("prof %s --tag tagName cpu.prof memory.prof block.prof mutex.prof", internal.MANUALCMD),
 		RunE: func(_ *cobra.Command, args []string) error {
 			return collector.RunCollector(args, tag)
 		},
@@ -74,12 +72,14 @@ func createProfAuto() *cobra.Command {
 	profileFlag := "profiles"
 	tagFlag := "tag"
 	countFlag := "count"
-	example := fmt.Sprintf(`prof %s --%s "BenchmarkGenPool" --%s "cpu,memory" --%s 10 --%s "tag1"`, shared.AUTOCMD, benchFlag, profileFlag, countFlag, tagFlag)
+	example := fmt.Sprintf(`prof %s --%s "BenchmarkGenPool" --%s "cpu,memory" --%s 10 --%s "tag1"`, internal.AUTOCMD, benchFlag, profileFlag, countFlag, tagFlag)
 
 	cmd := &cobra.Command{
-		Use:     shared.AUTOCMD,
-		Short:   "Wraps `go test` and `pprof` to benchmark code and gather profiling data for performance investigations.",
-		RunE:    runBenchmarks,
+		Use:   internal.AUTOCMD,
+		Short: "Wraps `go test` and `pprof` to benchmark code and gather profiling data for performance investigations.",
+		RunE: func(_ *cobra.Command, args []string) error {
+			return benchmark.RunBenchmarks(args)
+		},
 		Example: example,
 	}
 
@@ -119,11 +119,11 @@ func createTrackAutoCmd() *cobra.Command {
 	failFlag := "fail-on-regression"
 	thresholdFlag := "regression-threshold"
 	example := fmt.Sprintf(`prof track auto --%s "tag1" --%s "tag2" --%s "cpu" --%s "BenchmarkGenPool" --%s "summary"`, baseTagFlag, currentTagFlag, profileTypeFlag, benchNameFlag, outputFormatFlag)
-	longExplanation := fmt.Sprintf("This command only works if the %s command was used to collect and organize the benchmark and profile data, as it expects a specific directory structure generated by that process.", shared.AUTOCMD)
+	longExplanation := fmt.Sprintf("This command only works if the %s command was used to collect and organize the benchmark and profile data, as it expects a specific directory structure generated by that process.", internal.AUTOCMD)
 	shortExplanation := "If prof auto was used to collect the data, track auto can be used to analyze it, you just have to pass the tag name."
 
 	cmd := &cobra.Command{
-		Use:     shared.TrackAutoCMD,
+		Use:     internal.TrackAutoCMD,
 		Short:   shortExplanation,
 		Long:    longExplanation,
 		RunE:    runTrackAuto,
@@ -152,10 +152,10 @@ func createTrackManualCmd() *cobra.Command {
 	outputFormatFlag := "output-format"
 	failFlag := "fail-on-regression"
 	thresholdFlag := "regression-threshold"
-	example := fmt.Sprintf(`prof track %s --%s "path/to/profile_file.txt" --%s  "path/to/profile_file.txt"  --%s "summary"`, shared.TrackManualCMD, baseFlag, currentFlag, outputFormatFlag)
+	example := fmt.Sprintf(`prof track %s --%s "path/to/profile_file.txt" --%s  "path/to/profile_file.txt"  --%s "summary"`, internal.TrackManualCMD, baseFlag, currentFlag, outputFormatFlag)
 
 	cmd := &cobra.Command{
-		Use:     shared.TrackManualCMD,
+		Use:     internal.TrackManualCMD,
 		Short:   "Manually specify the paths to the profile text files you want to compare.",
 		RunE:    runTrackManual,
 		Example: example,
@@ -186,109 +186,9 @@ func createSetupCmd() *cobra.Command {
 	return cmd
 }
 
-func printConfiguration(benchArgs *args.BenchArgs, functionFilterPerBench map[string]config.FunctionFilter) {
-	slog.Info(
-		"Parsed arguments",
-		"Benchmarks", benchArgs.Benchmarks,
-		"Profiles", benchArgs.Profiles,
-		"Tag", benchArgs.Tag,
-		"Count", benchArgs.Count,
-	)
-
-	hasBenchFunctionFilters := len(functionFilterPerBench) > 0
-	if hasBenchFunctionFilters {
-		slog.Info("Benchmark Function Filter Configurations:")
-		for benchmark, cfg := range functionFilterPerBench {
-			slog.Info("Benchmark Config", "Benchmark", benchmark, "Prefixes", cfg.IncludePrefixes, "Ignore", cfg.IgnoreFunctions)
-		}
-	} else {
-		slog.Info("No benchmark configuration found in config file - analyzing all functions")
-	}
-}
-
-func runBenchAndGetProfiles(benchArgs *args.BenchArgs, benchmarkConfigs map[string]config.FunctionFilter) error {
-	slog.Info("Starting benchmark pipeline...")
-
-	var functionFilter config.FunctionFilter
-	globalFilter, hasGlobalFilter := benchmarkConfigs[shared.GlobalSign]
-	if hasGlobalFilter {
-		functionFilter = globalFilter
-	}
-
-	for _, benchmarkName := range benchArgs.Benchmarks {
-		slog.Info("Running benchmark", "Benchmark", benchmarkName)
-		if err := benchmark.RunBenchmark(benchmarkName, benchArgs.Profiles, benchArgs.Count, benchArgs.Tag); err != nil {
-			return fmt.Errorf("failed to run %s: %w", benchmarkName, err)
-		}
-
-		slog.Info("Processing profiles", "Benchmark", benchmarkName)
-		if err := benchmark.ProcessProfiles(benchmarkName, benchArgs.Profiles, benchArgs.Tag); err != nil {
-			return fmt.Errorf("failed to process profiles for %s: %w", benchmarkName, err)
-		}
-
-		slog.Info("Analyzing profile functions", "Benchmark", benchmarkName)
-
-		if !hasGlobalFilter {
-			functionFilter = benchmarkConfigs[benchmarkName]
-		}
-
-		args := &args.CollectionArgs{
-			Tag:             benchArgs.Tag,
-			Profiles:        benchArgs.Profiles,
-			BenchmarkName:   benchmarkName,
-			BenchmarkConfig: functionFilter,
-		}
-
-		if err := benchmark.CollectProfileFunctions(args); err != nil {
-			return fmt.Errorf("failed to analyze profile functions for %s: %w", benchmarkName, err)
-		}
-
-		slog.Info("Completed pipeline for benchmark", "Benchmark", benchmarkName)
-	}
-
-	slog.Info(shared.InfoCollectionSuccess)
-	return nil
-}
-
-func runBenchmarks(_ *cobra.Command, _ []string) error {
-	if len(benchmarks) == 0 {
-		return errors.New("benchmarks flag is empty")
-	}
-
-	if len(profiles) == 0 {
-		return errors.New("profiles flag is empty")
-	}
-
-	cfg, err := config.LoadFromFile(shared.ConfigFilename)
-	if err != nil {
-		slog.Info("No config file found at repository root; proceeding without function filters.", "expected", shared.ConfigFilename)
-		slog.Info("You can generate one with 'prof setup'. It will be placed at the root next to go.mod.")
-		cfg = &config.Config{}
-	}
-
-	if err = benchmark.SetupDirectories(tag, benchmarks, profiles); err != nil {
-		return fmt.Errorf("failed to setup directories: %w", err)
-	}
-
-	benchArgs := &args.BenchArgs{
-		Benchmarks: benchmarks,
-		Profiles:   profiles,
-		Count:      count,
-		Tag:        tag,
-	}
-
-	printConfiguration(benchArgs, cfg.FunctionFilter)
-
-	if err = runBenchAndGetProfiles(benchArgs, cfg.FunctionFilter); err != nil {
-		return err
-	}
-
-	return nil
-}
-
 // runSetup handles the setup command execution
 func runSetup(_ *cobra.Command, _ []string) error {
-	return config.CreateTemplate()
+	return internal.CreateTemplate()
 }
 
 var validFormats = map[string]bool{
