+++ internal/types.go
@@ -1,4 +1,6 @@
-package config
+package internal
+
+// #1 Config Fields
 
 // Config holds the main configuration for the prof tool.
 type Config struct {
@@ -16,4 +18,26 @@ type FunctionFilter struct {
 	// IgnoreFunctions ignores the function name after the last dot.
 	// Example: "Get,Set" excludes pool.Get() and cache.Set()
 	IgnoreFunctions []string `json:"ignore_functions,omitempty"`
+}
+
+// #2 - Function Arguments
+
+type LineFilterArgs struct {
+	ProfileFilters    map[int]float64
+	IgnoreFunctionSet map[string]struct{}
+	IgnorePrefixSet   map[string]struct{}
+}
+
+type CollectionArgs struct {
+	Tag             string
+	Profiles        []string
+	BenchmarkName   string
+	BenchmarkConfig FunctionFilter
+}
+
+type BenchArgs struct {
+	Benchmarks []string
+	Profiles   []string
+	Count      int
+	Tag        string
 }
